str to bytes conversion for Python 3
On Python 3 all the protocol communication must use bytes as strings are
"converted" to unicode. This is in theory a much cleaner implementation,
but poses some problems for code that should also run on Python 2.x.
This converts everything to bytes and adds the necessary compatibility
for 2.x where bytes is just str.